#77 - Refactoring: move logic to a separate hook
