+++ src/components/SimulationParameters.tsx
@@ -34,7 +34,7 @@ import AllIntermediateEvents from './interEvents/AllIntermediateEvents'
 import EventIcon from '@mui/icons-material/Event';
 import AllBatching from './batching/AllBatching';
 import DynamicFeedIcon from '@mui/icons-material/DynamicFeed';
-import { Dictionary } from './modelData';
+import useTabVisibility, {TABS} from './simulationParameters/useTabVisibility';
 
 const useStyles = makeStyles( (theme: Theme) => ({
     simParamsGrid: {
@@ -47,28 +47,6 @@ const useStyles = makeStyles( (theme: Theme) => ({
         "& .Mui-active": { color: theme.palette.info.dark }
     }
 }));
-
-enum TABS {
-    CASE_CREATION,
-    RESOURCE_CALENDARS,
-    RESOURCES,
-    RESOURCE_ALLOCATION,
-    BRANCHING_PROB,
-    INTERMEDIATE_EVENTS,
-    BATCHING,
-    SIMULATION_RESULTS
-}
-
-const tabs_name : { [key: string]: string } = {
-    CASE_CREATION: "Case Creation",
-    RESOURCE_CALENDARS: "Resource Calendars",
-    RESOURCES: "Resources",
-    RESOURCE_ALLOCATION: "Resource Allocation",
-    BRANCHING_PROB: "Branching Probabilities",
-    INTERMEDIATE_EVENTS: "Intermediate Events",
-    BATCHING: "Batching",
-    SIMULATION_RESULTS: "Simulation Results"
-};
 
 const tooltip_desc: { [key: string]: string } = {
     CASE_CREATION: 
@@ -136,41 +114,9 @@ const SimulationParameters = () => {
     const { formState: { errors, isValid, isSubmitted, submitCount }, getValues, handleSubmit } = formState
     const [ isScenarioParamsValid, setIsScenarioParamsValid ] = useState(true)
 
-    const [ visibleTabs, setVisibleTabs ] = useState(new Dictionary<string>())
-    const [ isEventsTabHidden, setIsEventsTabHidden ] = useState<boolean | undefined>(undefined)
+    const { visibleTabs } = useTabVisibility(eventsFromModel)
 
     const { onUploadNewModel } = useNewModel()
-
-    useEffect(() => {
-        if (eventsFromModel === undefined)
-            return
-
-        const areEventsEmpty = eventsFromModel.isEmpty()
-        if (areEventsEmpty !== isEventsTabHidden) {
-            setIsEventsTabHidden(areEventsEmpty)
-        }
-    }, [eventsFromModel])
-
-    useEffect(() => {
-        if (isEventsTabHidden === undefined || !visibleTabs.isEmpty()) {
-            // skip initialization if
-            // 1) no information yet loaded
-            // 2) we already have a final array
-            return 
-        }
-
-        const newVisibleTabs = Object.entries(tabs_name).reduce((acc: Dictionary<string>, [key, value]: any) => {
-            if (key === TABS[TABS.INTERMEDIATE_EVENTS] && isEventsTabHidden) {
-                // section is hidden, we don't add it to the general list
-                return acc
-            }
-
-            acc.add(key, value)
-            return acc
-        }, new Dictionary<string>())
-
-        setVisibleTabs(newVisibleTabs)
-    }, [isEventsTabHidden, visibleTabs])
 
     // validate both forms: scenatio params and json fields
     useEffect(() => {
